monkey-core/requirements.txt
ansible==2.10.0
ansible-base==2.10.1
ansible-runner==1.4.6
appdirs==1.4.4
attrs==19.3.0
autoflake==1.4
bcrypt==3.1.7
beautifulsoup4==4.9.1
black==20.8b1
boto==2.49.0
boto3==1.14.31
botocore==1.17.31
cached-property==1.5.1
cachetools==4.1.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cryptography==3.3.2
dirhash==0.2.1
distro==1.5.0
docker==4.2.1
docker-compose==1.26.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.15.2
flake8==3.8.3
Flask==1.1.2
google==2.0.3
google-api-core==1.17.0
google-api-python-client==1.8.3
google-auth==1.15.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.51.0
httplib2==0.19.0
idna==2.9
importlib-metadata==1.6.1
isort==5.5.3
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
jmespath==0.10.0
jsonschema==3.2.0
lockfile==0.12.2
MarkupSafe==1.1.1
mccabe==0.6.1
mongoengine==0.20.0
mypy==0.782
mypy-extensions==0.4.3
oauthlib==3.1.0
packaging==20.4
paramiko==2.7.1
parso==0.7.1
pathspec==0.8.0
pexpect==4.8.0
pkg-resources==0.0.0
pluggy==0.13.1
protobuf==3.14.0
psutil==5.7.0
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
pyls-black==0.4.6
pyls-isort==0.2.0
pyls-mypy==0.1.8
pymongo==3.10.1
PyNaCl==1.4.0
pyparsing==2.4.7
pyrsistent==0.16.0
python-daemon==2.2.4
python-dateutil==2.8.1
python-dotenv==0.13.0
python-jsonrpc-server==0.4.0
python-language-server==0.35.1
pytz==2020.1
PyYAML==5.3.1
regex==2020.7.14
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
s3transfer==0.3.3
scantree==0.0.1
six==1.15.0
soupsieve==2.0.1
termcolor==1.1.0
texttable==1.6.2
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
ujson==3.2.0
uritemplate==3.0.1
urllib3==1.25.9
websocket-client==0.57.0
Werkzeug==1.0.1
yapf==0.30.0
zipp==3.1.0
